Initial functionality testing.
Added a meta box for resources. Lists available collections. When collection is selected a tagcloud is shown. Clicking on a tag will show collection resource with that tag. There is also an optional checkbox to show untagged resources. A resource could be dragged and dropped onto the editor (works with both tinymce and QTags logics (a drop off point is holder for both). NB! Current version is for testing purposes only and is using a predefined test data. In addition no connection to the server can be established as of the moment.