Reworked the code a bit.
Moved service call functions into a standalone file. Added a functionality to replace links to uploaded resources with code that would download the resource (reusing the logic from KnowBrain; has the same limitations of working only with latest versions of Chrome and Opera). It might be a good idea to get it running using WP PHP side. Just create a page that would be passed some parameters and will just trigger the file download. This would require implementation of direct SSSREST calls from the server-side. This way we would not be exposing the credentials.