Added some methods to service impl.
The PHP-side implementation should be mostly working but would still need to be taken a look at and possibly simplified a lot. There should also be a logging possibility in case of some errors (automated one). Methods are still missing docstrings and not only. Also introduced new setting for specifying the REST service location.