Added initial version of raiting.
Using the JS API for the raiting of the post. Works for both single post view and multiple posts listing. Uses post ID embedded into an identifier of an article element to construct the URI. It would be a better idea to implement all of it on the server side and use local AJAX calls instead.