Service configuration detection.
Added a way to detect if service configuration settings are provided. In case they are left empty a notice will be displayed for the administrator. In addition the plugin will detect the situation of configuration not being provided and disable all the additional functionality. For the front-end it means that scripts will not be loaded. It also disables the additional metabox. One additional change for rating. In case of authenticated user the rating will be submitted using the user URI, otherwise an additional IP address will be added to fake a rating by unauthenticated user.