Code rewrite for new API (release 1.0.0-alpha.
Should be working, but has not been tested properly. The current server-side code seems to be having problems with providing a cumulated collection tags, thus the method of getting collections resources by tag was not properly tested. Also changed logging in case of service call error, the request data will also be logged if available. This would make it easier to determine calls to which methods have failed. Tested with the latest version of code from the middle point of today.