Added new logic of determining credentials to be used.
New logic is used when viewing posts. In case of a logged in user, service credentials of that user are tried to be found. If that is impossible, then the credentials of the current blog are being used. This is being done for submitting rating by currently logged in user. It would also affect the file download if the user is for some reason not allowed to access the file to be downloaded. In future, that kind of things should probably be done by using the backend service to make the calls (although that would add a synchronous call logic to the flow and probably make it slower). But the credentials could then be hidden from the anonymous users.