Changes to post view service calls.
The JS API is no longer used, thus the credentials of the user do not  get exposed. New solution is not finished yet, but should already be functional enough. New approach allows to use different credentials for different calls. File download and rating get is handled using blog credentials (the ones that belong to the creator of the posts). Rating submit is using logge in user credentials if possible and falls back to blog ones. Anonymous rating submission is no longer allowed.